perf: Wrote a quicker implementation of Iterate.
We don't need to check every cell for survivability - only live
cells and neighbors of live cells. This may actually be slower
for dense grids, but large, dense grids do not remain so for long,
and small, dense grids are going to be fast regardless. And this
implementation will be much faster for large, sparse grids.